refactor: topicDropDown 최초저장, 수정하기
